+++ apiExamples/basic.html
@@ -1,15 +1,7 @@
-<!-- icon variant -->
 <auro-button id="basicToastBtn">
   Show default notification
 </auro-button>
-<auro-toast style="display: block; margin: 0.5rem 0;" id="basicToast">
-  Default notification with no error type
-</auro-toast>
 
-<!-- no icon variant -->
-<auro-button id="basicToast-noIconBtn">
-  Show default notification with no icon
-</auro-button>
-<auro-toast style="display: block; margin: 0.5rem 0;" id="basicToast-noIcon" noIcon>
+<auro-toast style="display: block; margin: 0.5rem 0;" id="basicToast">
   Default notification with no error type
 </auro-toast>
