+++ apiExamples/custom.html
@@ -1,0 +1,7 @@
+<auro-button id="customToastBtn">
+  Show default notification
+</auro-button>
+
+<custom-toast style="display: block; margin: 0.5rem 0;" id="customToast">
+  Default notification with no error type
+</custom-toast>
