+++ apiExamples/disable-auto-hide.html
@@ -1,0 +1,7 @@
+<auro-button id="disableHideToastBtn">
+  Show default notification
+</auro-button>
+
+<auro-toast style="display: block; margin: 0.5rem 0;" id="disableHideToast">
+  Default notification with no error type
+</auro-toast>
