--- apiExamples/error.html
@@ -1,4 +1,0 @@
-<auro-button id="errorToastBtn">Show error notification</auro-button>
-<auro-toast style="display: block; margin: 0.5rem 0;" variant="error" id="errorToast"> Unable to add lap infant. Please try again  </auro-toast>
-<auro-button id="errorToast-noIconBtn">Show error notification with no icon</auro-button>
-<auro-toast style="display: block; margin: 0.5rem 0;" variant="error" id="errorToast-noIcon" noIcon> Unable to add lap infant. Please try again  </auro-toast>
