+++ apiExamples/multiple-toasts.html
