+++ apiExamples/multiple-toasts.js
