+++ apiExamples/no-icon.js
