--- apiExamples/showToast.js
@@ -1,7 +1,0 @@
-export function showToast(toastID) {
-  const toast = document.querySelector(toastID);
-
-  if (!toast.hasAttribute("visible")) {
-    toast.setAttribute("visible", true);
-  }
-}
