--- apiExamples/success.html
@@ -1,4 +1,0 @@
-<auro-button id="successToastBtn">Show success toast</auro-button>
-<auro-toast style="display: block; margin: 0.5rem 0;" variant="success" id="successToast"> Successfully added lap infant  </auro-toast>
-<auro-button id="successToast-noIconBtn">Show success toast with no icon</auro-button>
-<auro-toast style="display: block; margin: 0.5rem 0;" variant="success" id="successToast-noIcon" noIcon> Successfully added lap infant  </auro-toast>
