+++ apiExamples/time-til-hide.js
