+++ package.json
@@ -21,17 +21,17 @@
     "node": ">=20"
   },
   "dependencies": {
-    "lit": "^3.3.1"
+    "lit": "^3.3.2"
   },
   "devDependencies": {
-    "@alaskaairux/icons": "^5.11.0",
-    "@aurodesignsystem/auro-button": "^12.2.0",
-    "@aurodesignsystem/auro-icon": "^9.1.0",
-    "@aurodesignsystem/auro-cli": "^3.0.4",
+    "@alaskaairux/icons": "^5.14.0",
+    "@aurodesignsystem/auro-button": "^12.3.0",
+    "@aurodesignsystem/auro-cli": "^3.5.0",
     "@aurodesignsystem/auro-config": "^1.3.1",
+    "@aurodesignsystem/auro-icon": "^9.1.1",
     "@aurodesignsystem/auro-library": "^5.5.4",
-    "@aurodesignsystem/design-tokens": "^8.5.0",
-    "@aurodesignsystem/webcorestylesheets": "^10.1.1",
+    "@aurodesignsystem/design-tokens": "^8.7.0",
+    "@aurodesignsystem/webcorestylesheets": "^10.1.4",
     "husky": "^9.1.7"
   },
   "browserslist": [
@@ -66,7 +66,12 @@
   "exports": {
     "./package.json": "./package.json",
     "./readme.md": "./README.md",
-    ".": "./dist/registered.js",
+    "./custom-elements.json": "./custom-elements.json",
+    ".": {
+      "module": "./dist/registered.js",
+      "types": "./dist/index.d.ts",
+      "default": "./dist/registered.js"
+    },
     "./demo/*.md": "./demo/*.md",
     "./demo/*.js": "./demo/*.min.js",
     "./class": {
@@ -82,5 +87,7 @@
     "README.md",
     "LICENSE",
     "NOTICE"
-  ]
+  ],
+  "custom-elements": "./custom-elements.json",
+  "customElements": "custom-elements.json"
 }
